covidtask/requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.12.7
distlib==0.3.6
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.2
pipenv==2022.11.30
platformdirs==2.6.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
virtualenv==20.17.1
virtualenv-clone==0.5.7
